mcp/client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "deepseek-sdk>=0.1.0",
    "mcp>=1.1.1",
    "openai>=1.93.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.4",
]
